feat: support decimal seconds for macro step sleep duration
Change sleep field from int to double to allow fractional seconds (e.g. 0.5s).
Update UI input, JSON parsing, Task.Delay conversion, and countdown display accordingly.

+++ FFXIVMacroController/Model/MacroModel.cs
@@ -11,7 +11,7 @@ namespace FFXIVMacroControllerApp.Model
     {
         public Keys key { get; set; }
         public int keyNumber { get; set; }
-        public int sleep { get; set; }
+        public double sleep { get; set; }
         public Types type { get; set; }
         public int typeNumber { get; set; }
         public string inputText { get; set; }
